[FEAT] LoginScreen 통합 및 string resource 영어, 한글 대응
