+++ feature/login/src/main/kotlin/com/guesthouse/login/ui/route/LoginRoute.kt
@@ -1,11 +1,75 @@
 package com.guesthouse.login.ui.route
 
+import android.content.Context
+import android.util.Log
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.ui.platform.LocalContext
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.guesthouse.base.use
 import com.guesthouse.login.ui.screen.LoginScreen
+import com.guesthouse.login.ui.viewmodel.LoginContract
+import com.guesthouse.login.ui.viewmodel.LoginViewModel
+import com.kakao.sdk.auth.model.OAuthToken
+import com.kakao.sdk.common.model.ClientError
+import com.kakao.sdk.common.model.ClientErrorCause
+import com.kakao.sdk.user.UserApiClient
+import kotlinx.coroutines.flow.collectLatest
 
 @Composable
 fun LoginRoute(
+    viewModel: LoginViewModel = hiltViewModel(),
     onEmailLoginClick:() -> Unit
 ) {
-    LoginScreen(onEmailLoginClick)
+    val context = LocalContext.current
+    val (state, event, effect) = use(viewModel)
+
+    LaunchedEffect(key1 = effect) {
+        effect.collectLatest { effect ->
+            when (effect) {
+                LoginContract.Effect.KakaoLogin -> kakaoLogin(context)
+            }
+        }
+    }
+    LoginScreen(
+        state = state,
+        event = event,
+        onEmailLoginClick = onEmailLoginClick
+    )
+}
+
+private fun kakaoLogin(context: Context) {
+
+    // 카카오계정으로 로그인 공통 callback 구성
+    // 카카오톡으로 로그인 할 수 없어 카카오계정으로 로그인할 경우 사용됨
+    val callback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
+        if (error != null) {
+            Log.e("LOGIN", "카카오계정으로 로그인 실패", error)
+        } else if (token != null) {
+            Log.i("LOGIN", "카카오계정으로 로그인 성공 ${token.accessToken}")
+        }
+    }
+
+    // 카카오톡이 설치되어 있으면 카카오톡으로 로그인, 아니면 카카오계정으로 로그인
+    if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
+        UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
+            if (error != null) {
+                Log.e("LOGIN", "카카오톡으로 로그인 실패", error)
+
+                // 사용자가 카카오톡 설치 후 디바이스 권한 요청 화면에서 로그인을 취소한 경우,
+                // 의도적인 로그인 취소로 보고 카카오계정으로 로그인 시도 없이 로그인 취소로 처리 (예: 뒤로 가기)
+                if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
+                    return@loginWithKakaoTalk
+                }
+
+                // 카카오톡에 연결된 카카오계정이 없는 경우, 카카오계정으로 로그인 시도
+                UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
+            } else if (token != null) {
+                Log.i("LOGIN", "카카오톡으로 로그인 성공 ${token.accessToken}")
+            }
+        }
+    } else {
+        UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
+    }
+
 }
