+++ feature/login/src/main/res/values-ko/strings.xml
@@ -9,7 +9,7 @@
     <string name="email_login_button">이메일 로그인 버튼</string>
     <string name="main_logo">메인로고</string>
     <string name="icon">아이콘</string>
-    <string name="business_or_admin_sign_up">비즈니스/관리자 회원가입 ></string>
     <string name="email_login">이메일 로그인</string>
     <string name="close_button">닫기 버튼</string>
+    <string name="business_admin_sign_up">비즈니스 / 관리자 회원가입</string>
 </resources>
