+++ feature/login/src/main/res/values/strings.xml
@@ -9,7 +9,7 @@
     <string name="email_login_button">Email login button</string>
     <string name="main_logo">mainLogo</string>
     <string name="icon">icon</string>
-    <string name="business_or_admin_sign_up">businessOrAdminSignUp</string>
     <string name="email_login">emailLogin</string>
     <string name="close_button">closeButton</string>
+    <string name="business_admin_sign_up">Business / Admin SignUp</string>
 </resources>
